Refactor 'value' to 'byValue' in methods and filenames for Size
