+++ lib/assets/javascripts/builder/components/form-components/_constants/_fill.js
@@ -11,6 +11,11 @@ module.exports = {
       max: 45,
       step: 0.5
     },
+    DefaultInput100: {
+      MIN: 0,
+      MAX: 100,
+      STEP: 1
+    },
     DEFAULT_RANGE: [ 5, 20 ]
   },
