+++ lib/assets/javascripts/builder/components/form-components/editors/size/size-by-value-content-view.js
@@ -1,6 +1,6 @@
 var Backbone = require('backbone');
 var CoreView = require('backbone/core-view');
-var template = require('./size-value-content-view.tpl');
+var template = require('./size-by-value-content-view.tpl');
 
 module.exports = CoreView.extend({
   events: {
