+++ lib/assets/javascripts/builder/components/form-components/editors/size/size-by-value-view.js
@@ -2,17 +2,19 @@ var _ = require('underscore');
 var Backbone = require('backbone');
 var CoreView = require('backbone/core-view');
 
-var template = require('./size-value-view.tpl');
+var template = require('./size-by-value-view.tpl');
 var DialogModel = require('builder/components/dialog/dialog-model');
 var DialogView = require('builder/components/dialog/dialog-view');
 
 var StackLayoutView = require('builder/components/stack-layout/stack-layout-view');
 var ColumnListView = require('builder/components/custom-list/column-list/column-list-view');
 var columnListQuantificationMethodItemTemplate = require('builder/components/custom-list/column-list/column-list-quantification-method-item.tpl');
-var SizeValueContentView = require('./size-value-content-view');
+var SizeByValueContentView = require('./size-by-value-content-view');
 
 var PopupManager = require('builder/components/popup-manager');
+
 var FillConstants = require('builder/components/form-components/_constants/_fill');
+var DialogConstants = require('builder/components/form-components/_constants/_dialogs');
 
 var checkAndBuildOpts = require('builder/helpers/required-opts');
 
@@ -139,7 +141,7 @@ module.exports = CoreView.extend({
     this._dialogView.render();
     this._dialogView.show();
 
-    this._popupManager.append('float');
+    this._popupManager.append(DialogConstants.Mode.FLOAT);
     this._popupManager.track();
   },
 
@@ -180,7 +182,7 @@ module.exports = CoreView.extend({
   },
 
   _createSizeValueContentView: function (stackLayoutModel, opts) {
-    var view = new SizeValueContentView({
+    var view = new SizeByValueContentView({
       stackLayoutModel: stackLayoutModel,
       model: this.model,
       min: this.options.min,
