+++ lib/assets/javascripts/builder/components/form-components/editors/size/size-by-value-view.tpl
