+++ lib/assets/javascripts/builder/components/form-components/editors/size/size.js
@@ -7,13 +7,9 @@ var createRadioLabelsTabPane = require('builder/components/tab-pane/create-radio
 var EditorHelpers = require('builder/components/form-components/editors/editor-helpers-extend');
 
 var SizeFixedView = require('./size-fixed-view');
-var SizeValueView = require('./size-value-view');
+var SizeByValueView = require('./size-by-value-view');
 
 var FillConstants = require('builder/components/form-components/_constants/_fill');
-
-var DEFAULT_INPUT_MIN_VALUE = 0;
-var DEFAULT_INPUT_MAX_VALUE = 100;
-var DEFAULT_INPUT_STEP_VALUE = 1;
 
 Backbone.Form.editors.Size = Backbone.Form.editors.Base.extend({
   className: 'Form-InputSize CDB-Text',
@@ -83,9 +79,9 @@ Backbone.Form.editors.Size = Backbone.Form.editors.Base.extend({
     var attrs = this.options.editorAttrs;
     if (attrs && attrs.hidePanes) {
       if (!_.contains(attrs.hidePanes, FillConstants.Panes.FIXED)) tabPaneTabs.push(this._buildFixedPane());
-      if (!_.contains(attrs.hidePanes, FillConstants.Panes.VALUE)) tabPaneTabs.push(this._buildValuePane());
+      if (!_.contains(attrs.hidePanes, FillConstants.Panes.BY_VALUE)) tabPaneTabs.push(this._buildByValuePane());
     } else {
-      tabPaneTabs = [this._buildFixedPane(), this._buildValuePane()];
+      tabPaneTabs = [this._buildFixedPane(), this._buildByValuePane()];
     }
     return tabPaneTabs;
   },
@@ -124,13 +120,13 @@ Backbone.Form.editors.Size = Backbone.Form.editors.Base.extend({
     return fixedPane;
   },
 
-  _buildValuePane: function () {
+  _buildByValuePane: function () {
     var self = this;
     var valuePane = {
-      name: FillConstants.Panes.VALUE,
-      label: _t('form-components.editors.fill.input-number.value'),
+      name: FillConstants.Panes.BY_VALUE,
+      label: _t('form-components.editors.fill.input-number.by-value'),
       createContentView: function () {
-        return self._generateValueContentView();
+        return self._generateByValueContentView();
       }
     };
     return valuePane;
@@ -186,20 +182,20 @@ Backbone.Form.editors.Size = Backbone.Form.editors.Base.extend({
     var editorAttrs = this.options.editorAttrs;
     this._fixedView = new SizeFixedView({
       model: this._sizeModel,
-      min: editorAttrs && editorAttrs.min || DEFAULT_INPUT_MIN_VALUE,
-      max: editorAttrs && editorAttrs.max || DEFAULT_INPUT_MAX_VALUE,
-      step: editorAttrs && editorAttrs.step || DEFAULT_INPUT_STEP_VALUE
+      min: editorAttrs && editorAttrs.min || FillConstants.Size.DefaultInput100.MIN,
+      max: editorAttrs && editorAttrs.max || FillConstants.Size.DefaultInput100.MAX,
+      step: editorAttrs && editorAttrs.step || FillConstants.Size.DefaultInput100.STEP
     });
     return this._fixedView;
   },
 
-  _generateValueContentView: function () {
+  _generateByValueContentView: function () {
     var editorAttrs = this.options.editorAttrs;
-    this._valueView = new SizeValueView({
+    this._valueView = new SizeByValueView({
       model: this._sizeModel,
       columns: this._columns,
-      min: editorAttrs && editorAttrs.min || DEFAULT_INPUT_MIN_VALUE,
-      max: editorAttrs && editorAttrs.max || DEFAULT_INPUT_MAX_VALUE,
+      min: editorAttrs && editorAttrs.min || FillConstants.Size.DefaultInput100.MIN,
+      max: editorAttrs && editorAttrs.max || FillConstants.Size.DefaultInput100.MAX,
       popupConfig: {
         cid: this.cid,
         $el: this.$el
